rename components with uppercase first letter
